api/requirements.txt
Django==3.2.7
djangorestframework==3.12.4
psycopg2-binary==2.9.1
gunicorn==20.1.0
pandas==1.3.3
pytest==6.2.5
